refactor: adminのネームスペースを作成
